requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
streamlit>=1.28.0
Pillow>=9.0.0
numpy>=1.24.0
ftfy>=6.0.0
regex>=2023.0.0
git+https://github.com/openai/CLIP.git
requests>=2.28.0
pandas>=2.0.0
PyYAML>=6.0
huggingface_hub>=0.19.0
inference-sdk>=0.9.0
transformers>=4.40.0
python-dotenv>=1.0.1
# SAM segmentation support:
git+https://github.com/facebookresearch/segment-anything.git
